requirements.txt
asgiref==3.6.0
autopep8==2.0.1
beautifulsoup4==4.11.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.4
cssselect==1.2.0
defusedxml==0.7.1
dj-database-url==1.2.0
dj-rest-auth==2.2.5
Django==3.2.15
django-allauth==0.52.0
django-autoslug==1.9.8
django-cors-headers==3.13.0
django-filter==22.1
django-quill-editor==0.1.40
djangorestframework==3.14.0
djangorestframework-jsonapi==6.0.0
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
lxml==4.9.2
markdown2==2.4.6
oauthlib==3.2.2
Pillow==9.3.0
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.6.0
pyquery==2.0.0
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.7
readtime==2.0.0
requests==2.28.1
requests-oauthlib==1.3.1
soupsieve==2.3.2.post1
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.26.13
whitenoise==6.2.0
